+++ test/Recore/OfTests.cs
@@ -1,4 +1,3 @@
-using System.Text.Json;
 using Xunit;
 
 namespace Recore.Tests
@@ -7,29 +6,8 @@ namespace Recore.Tests
     {
         class Address : Of<string>
         {
-            public Address() { }
             public Address(string value) => Value = value;
         }
-
-        [OfJson(typeof(JsonAddress), typeof(string))]
-        class JsonAddress : Of<string>
-        {
-            public JsonAddress() { }
-            public JsonAddress(string value) => Value = value;
-        }
-
-        class House
-        {
-            public JsonAddress Street { get; set; }
-        }
-
-        class Person
-        {
-            public string Name { get; set; }
-            public int Age { get; set; }
-        }
-
-        class User : Of<Person> { }
 
         [Fact]
         public void Equals_()
@@ -49,67 +27,6 @@ namespace Recore.Tests
             Address address2 = null;
 
             Assert.NotEqual(address1, address2);
-        }
-
-        [Fact]
-        public void ToJson()
-        {
-            Assert.Equal(
-                expected: "{\"Value\":\"1 Microsoft Way\"}",
-                actual: JsonSerializer.Serialize(new Address("1 Microsoft Way")));
-
-            Assert.Equal(
-                expected: "\"1 Microsoft Way\"",
-                actual: JsonSerializer.Serialize(new JsonAddress("1 Microsoft Way")));
-
-            Assert.Equal(
-                expected: "{\"Street\":\"123 Main St\"}",
-                actual: JsonSerializer.Serialize(new House
-                {
-                    Street = new JsonAddress("123 Main St")
-                }));
-
-            Assert.Equal(
-                expected: "{\"Name\":\"Mario\",\"Age\":42}",
-                actual: JsonSerializer.Serialize(new User
-                {
-                    Value = new Person { Name = "Mario", Age = 42 }
-                }));
-        }
-
-        [Fact]
-        public void FromJson()
-        {
-            // This throws because `Address` does not have `[OfJson(...)]`
-            Assert.Throws<JsonException>(
-                () => JsonSerializer.Deserialize<Address>("\"1 Microsoft Way\""));
-
-            Assert.Equal(
-                expected: new Address("1 Microsoft Way"),
-                actual: JsonSerializer.Deserialize<Address>("{\"Value\":\"1 Microsoft Way\"}"));
-
-            Assert.Equal(
-                expected: new JsonAddress("1 Microsoft Way"),
-                actual: JsonSerializer
-                    .Deserialize<Of<string>>("\"1 Microsoft Way\"")
-                    .To<JsonAddress>());
-
-            Assert.Equal(
-                expected: new JsonAddress("1 Microsoft Way"),
-                actual: JsonSerializer.Deserialize<JsonAddress>("\"1 Microsoft Way\""));
-
-            Assert.Equal(
-                expected: new JsonAddress("123 Main St"),
-                actual: JsonSerializer.Deserialize<House>("{\"Street\":\"123 Main St\"}").Street);
-
-            var deserializedUser = JsonSerializer.Deserialize<User>("{\"Name\":\"Mario\",\"Age\":42}");
-            Assert.Equal(
-                expected: "Mario",
-                actual: deserializedUser.Value.Name);
-
-            Assert.Equal(
-                expected: 42,
-                actual: deserializedUser.Value.Age);
         }
     }
 }
